Feature/dapp (#79)
* add box mainnet logic contract

* update environment variables

* update dependencies

* factor out VRF interfaces

* Add Seed Storage contract and test stub

* add order oracle

* add pizza interfaces

* add initial deployment script

* add initial pizza contract (WIP)

* Update Deployment scripts

remove env vars that are not actually used.  Add a mint pizza sequence diagram.  update docs, etc

* Add fulfillment event

update fallback function

* npm upgrade

* dont run the bonding curve tests all the time

they take a while

* moar tool updates

* clean up seed storage deployment

add a deployment to mumbai

* update gitignore and packages

* rename mumbai to maticmum

when ethers added official support, they chose maticmum as the name.

see here: https://github.com/ethers-io/ethers.js/issues/1546#issuecomment-885956859

* add deployment for polygon random consumer

* pizza storage set random consumer

* update docs

* Remove purchase logic, use box token id

remove the logic to purchase a pizza.  it now just redirects to the box contract (for api consistency).  Also use the box id as the token id for the pizza for now.

* missed interface upgrade

* clean up the order api consumer deployment

add the job as a string and use assembly to load it into bytes32. note this is dangerous.

* check the chainlink callback is a contract

use openzep's _msgSender() in some places

* remove references to kovan

update packages

* fix test

* reanme

* update pizza deployment

* Fix the IPFS hash resolution

The IPFS hashes are too big when coming in from the chainlink oracle to fit in bytes32 so they get converted from base58 into a uint and the multihash bytes are stripped off, then recombined when reading out the hash when calling tokenURI.

* PR self review

* add a couple tests and scripts

* add missing interface members

add missing function parameters add a few missing checks. change the data type of the address to a uint

* add toggle for sale

* add partially implemented dapp

* Changes to align old and new dapp patterns, some renaming, and tried to align the semantics of the contracts.

* further stub out the dapp

* stub out pricing and token fetch

* add webpack and compile

* couple deployments remove unnecessary methods update dapp

* update docs

Co-authored-by: Julian Bleecker <bleeckerj@gmail.com>